+++ src/main/java/org/sitmun/authorization/client/mapper/ApplicationMapper.java
@@ -3,6 +3,7 @@ package org.sitmun.authorization.client.mapper;
 import java.util.List;
 import org.mapstruct.AfterMapping;
 import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
 import org.mapstruct.ReportingPolicy;
 import org.sitmun.authorization.client.dto.ApplicationDtoLittle;
@@ -12,6 +13,7 @@ import org.sitmun.domain.user.User;
 @Mapper(unmappedTargetPolicy = ReportingPolicy.WARN)
 public interface ApplicationMapper {
 
+  @Mapping(source = "creator", target = "pointOfContact")
   ApplicationDtoLittle map(Application application);
 
   List<ApplicationDtoLittle> map(List<Application> applications);