+++ src/test/java/org/sitmun/authorization/client/dto/ApplicationDtoLittleTest.java
@@ -32,7 +32,7 @@ class ApplicationDtoLittleTest {
     dto.setIsUnavailable(false);
     dto.setAppPrivate(false);
     dto.setLastUpdate(testDate);
-    dto.setCreator("Test Creator");
+    dto.setPointOfContact("contact@example.com");
     dto.setHeaderParams(testHeaderParams);
     dto.setConfig(testConfig);
     dto.setExternalUrl("https://www.idee.es");
@@ -47,7 +47,7 @@ class ApplicationDtoLittleTest {
     assertThat(dto.getIsUnavailable()).isFalse();
     assertThat(dto.getAppPrivate()).isFalse();
     assertThat(dto.getLastUpdate()).isEqualTo(testDate);
-    assertThat(dto.getCreator()).isEqualTo("Test Creator");
+    assertThat(dto.getPointOfContact()).isEqualTo("contact@example.com");
     assertThat(dto.getHeaderParams()).isEqualTo(testHeaderParams);
     assertThat(dto.getConfig()).isEqualTo(testConfig);
     assertThat(dto.getExternalUrl()).isEqualTo("https://www.idee.es");